# PERF: __contains__ method for Categorical