--- a/pandas/core/indexes/category.py
+++ b/pandas/core/indexes/category.py
@@ -323,20 +323,10 @@ class CategoricalIndex(Index, accessor.PandasDelegate):
 
     @Appender(_index_shared_docs['__contains__'] % _index_doc_kwargs)
     def __contains__(self, key):
-        hash(key)
-
-        if self.categories._defer_to_indexing:
-            return key in self.categories
-
         return key in self.values
 
     @Appender(_index_shared_docs['contains'] % _index_doc_kwargs)
     def contains(self, key):
-        hash(key)
-
-        if self.categories._defer_to_indexing:
-            return self.categories.contains(key)
-
         return key in self.values
 
     def __array__(self, dtype=None):