@topper-123 2018-05-13
Nice.

wrt #20395: Now I get test result of 4.1 ms as compared 13.8 ms beforehand. So while this is much better, the test case of the ``pd.Index`` is at 125 µs, so #20395 is not completely solved...