@topper-123 2018-05-13
Even faster is

```python
def __contains__(self, key):
    """Returns True if `key` is in this Categorical."""
    code_values = self.categories.get_indexer([key])
    return code_values[0] in self._engine
```

Your code: 658 µs
My code: 266 ns