@topper-123 2018-05-13
Your code was good, though ise ``get_loc`` rather than ``get_indexer`` as this is faster.

A related improvement would be to look at ``CategoricalIndex.__contains__``. There the line 

```python
    return key in self.values
```
can be replaced with 

```python
        code_value = self.categories.get_loc(key)
        return code_value in self._engine
```

to gain some additional speedup:

```python
>>> n = 1_000_000
>>> c = pd.Categorical(list('a' * n + 'b' * n + 'c' * n))  # using get_indexer
>>> %timeit 'b' in c
556 µs ± 12.5 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
>>> ci = pd.CategoricalIndex(list('a' * n + 'b' * n + 'c' * n))
>>> %timeit 'b' in ci
3,11 µs ± 19.6 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

Also in ``CategoricalIndex.contains`` you can replace ``return key in self.values`` with ``return key in self``.