@jorisvandenbossche 2018-06-04
> , though use get_loc rather than get_indexer as this is faster.

General question, wouldn't simply `val in self.categories` not even be faster and simpler?