@topper-123 2018-06-04
@jorisvandenbossche: in general no, as some categories might not be present in the `´Categorical``/``categoricalIndex``...