@TomAugspurger 2018-06-04
> as some categories might not be present in the 

How is that possible? The only one I can think of is NA, so the check could be something like `val in self.categories or isna(val) and self.isna().any()`