@topper-123 2018-06-04
Maybe I wasn't communicating clearly, sorry about that. What I meant was this:

```python
>>> c = pd.CategoricalIndex('a a'.split(), categories='a b'.split())
>>> 'b' in c
False
>>< 'b' in c.categories
True
```

So to check for membership in categories can give different results than checking membership in the values. Was this not the question from @jorisvandenbossche, or did I misunderstand?