@TomAugspurger 2018-06-04
Oh, of course.

Though I wonder if checking `val in self.categories` is still worthwhile to get a quick no / maybe.